Done moving files, reached final folder structure, tests works
