+++ kenobi/helpers/emulator.py
@@ -8,9 +8,8 @@ from os import system
 from playsound import playsound
 from pynput.keyboard import Controller, Key
 
-from operating_system import OperatingSystem
-
-from custom_logger import CustomLogger
+from .operating_system import OperatingSystem
+from .custom_logger import CustomLogger
 
 
 class Emulator:
