+++ kenobi/helpers/websocket_server.py
@@ -6,9 +6,9 @@ import asyncio
 
 import websockets
 
-from custom_logger import CustomLogger
-from emulator import Emulator
-from message_parser import MessageParser as mp
+from .custom_logger import CustomLogger
+from .emulator import Emulator
+from .message_parser import MessageParser as mp
 
 HOST = "0.0.0.0"
 PORT = 9999
